+++ docs/core/concepts/secrets.md
@@ -33,10 +33,11 @@ For example, given an environment with the environment variable `PREFECT__CONTEX
 ```
 :::
 ::: tab Secret API
+
 ```python
->>> from prefect.client.secrets import Secret
->>> s = Secret("FOO")
->>> s.get()
+>> > from prefect.backend.secrets import Secret
+>> > s = Secret("FOO")
+>> > s.get()
 'mypassword'
 ```
 :::
