+++ src/prefect/backend/__init__.py
@@ -1,3 +1,5 @@
+from prefect.backend.client import Client
+from prefect.backend.secrets import Secret
 from prefect.backend.task_run import TaskRunView
 from prefect.backend.flow_run import FlowRunView, execute_flow_run
 from prefect.backend.flow import FlowView
