+++ src/prefect/cli/auth.py
@@ -2,7 +2,8 @@ import click
 from click.exceptions import Abort
 from tabulate import tabulate
 
-from prefect import Client, config
+from prefect import config
+from prefect.backend.client import Client
 from prefect.utilities.exceptions import AuthorizationError, ClientError
