+++ src/prefect/cli/get.py
@@ -2,7 +2,7 @@ import click
 import pendulum
 from tabulate import tabulate
 
-from prefect.client import Client
+from prefect.backend import Client
 from prefect.utilities.graphql import EnumValue, with_args
