+++ src/prefect/cli/heartbeat.py
@@ -3,7 +3,7 @@ import time
 import click
 
 from prefect import config
-from prefect.client import Client
+from prefect.backend import Client
 
 
 @click.group(hidden=True)
