--- src/prefect/client/__init__.py
@@ -1,2 +1,0 @@
-from prefect.client.client import Client
-from prefect.client.secrets import Secret
