+++ src/prefect/engine/cloud/task_runner.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, Dict, Iterable, Optional, Tuple
 import pendulum
 
 import prefect
-from prefect.client import Client
+from prefect.backend import Client
 from prefect.core import Edge, Task
 from prefect.engine.result import Result
 from prefect.engine.runner import ENDRUN, call_state_handlers
