+++ src/prefect/storage/gitlab.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, Any
 from urllib.parse import quote_plus
 
 import prefect
-from prefect.client import Secret
+from prefect.backend import Secret
 from prefect.storage import Storage
 from prefect.utilities.storage import extract_flow_from_file
