+++ src/prefect/tasks/dropbox/dropbox.py
@@ -2,7 +2,7 @@ import warnings
 
 import dropbox
 
-from prefect.client import Secret
+from prefect.backend import Secret
 from prefect.core import Task
 from prefect.utilities.tasks import defaults_from_attrs
