+++ src/prefect/tasks/gcp/storage.py
@@ -7,7 +7,7 @@ from time import sleep
 from google.cloud.exceptions import NotFound
 
 from prefect import context
-from prefect.client import Secret
+from prefect.backend import Secret
 from prefect.core import Task
 from prefect.utilities.gcp import get_storage_client
 from prefect.utilities.tasks import defaults_from_attrs
