+++ src/prefect/tasks/jira/jira_service_desk.py
@@ -1,5 +1,5 @@
 from prefect import Task
-from prefect.client import Secret
+from prefect.backend import Secret
 from prefect.utilities.tasks import defaults_from_attrs
 from typing import cast
 from typing import Any
