+++ src/prefect/tasks/redis/redis_tasks.py
@@ -1,7 +1,7 @@
 import redis
 
 from prefect import Task
-from prefect.client import Secret
+from prefect.backend import Secret
 from prefect.utilities.tasks import defaults_from_attrs
