+++ src/prefect/tasks/secrets/base.py
@@ -1,6 +1,6 @@
 import warnings
 
-from prefect.client.secrets import Secret as _Secret
+from prefect.backend.secrets import Secret as _Secret
 from prefect.core.task import Task
 from prefect.engine.results import SecretResult
